% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{approximate_color}
\alias{approximate_color}
\title{Approximate Color From Complex Styles}
\usage{
approximate_color(color, url)
}
\arguments{
\item{color}{character(1L) a value used as the "fill" attribute of an SVG
element.}

\item{url}{"url-data" object, typically kept as the "url" attribute of
"svg_chopped" objects.}
}
\value{
character(1L), a hex color, or NA_character_ if the fill could not
be approximated by a color.
}
\description{
The "fill" and "stroke" attributes to SVG elements may be specified in the
form "url(#id)" where "id" is the DOM id of another SVG element.  This is
used to implement complex colors such as gradients and patterns.  This
function will attempt to represent the complex fills with a single color if
it can based on data from the url-referenced object.
}
\details{
Currently only gradients are approximated.  They are approximated by taking
the arithmetic mean of the stop color RGB values.  The mean stop opacity is
attached as an attribute.
}
\examples{
svg <- chop(R_logo())
## str(svg) to see structure
fill.1 <- attr(svg[[2]], 'style-computed')[['fill']]
fill.1 # A gradient fill
approximate_color(fill.1, attr(svg, 'url'))
}
\seealso{
\code{\link[=chop]{chop()}}, \code{\link[=get_fills]{get_fills()}} and \code{\link[=get_strokes]{get_strokes()}} for easily retrieving
approximated gradient fills and strokes.
}
