% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{approximate_fill}
\alias{approximate_fill}
\title{Approximate Fill}
\usage{
approximate_fill(fill, url)
}
\arguments{
\item{fill}{character(1L) a value used as the "fill" attribute of an SVG
element.}

\item{url}{"url-data" object, typically kept as the "url" attribute of
"svg_chopped_list" objects.}
}
\value{
character(1L), a hex color, or NA_character_ if the fill could not
be approximated by a color.
}
\description{
The "fill" attribute to SVG elements may be specified in the form "url(#id)"
where "id" is the DOM id of another SVG element.  This is used to implement
complex fills such as gradients and patterns.  This function will attempt to
represent the complex fills with a single color if it can based on data from
the url-referenced object.
}
\details{
Currently only gradients are approximated.  They are approximated by taking
the arithmetic mean of the stop color RGB values.
}
\examples{
svg <- process_svg(file.path(R.home(), 'doc', 'html', 'Rlogo.svg'))
fill.1 <- attr(svg[[1]][[2]], 'style-computed')[['fill']]
fill.1 # A gradient fill
approximate_fill(fill.1, attr(svg, 'url'))
}
\seealso{
\code{\link[=process_svg]{process_svg()}}
}
