% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.svg_chopped_list}
\alias{as.svg_chopped_list}
\title{Convert Objects into "svg_chopped" Objects into "svg_chopped_list"}
\usage{
as.svg_chopped_list(x)
}
\arguments{
\item{x}{a list of "svg_chopped" objects}
}
\value{
an "svg_chopped_list" object
}
\description{
Useful in cases where we wish to take a list of "svg_chopped" objects, e.g.
because we produced them with \code{lapply}, but wish to have access to the
"svg_chopped_list" methods.
}
\examples{
svgs <- lapply(svg_samples()[1:2], chop)
\dontrun{
plot(svgs)  ## Error!
}
\donttest{
plot(as.svg_chopped_list(svgs), mfrow=c(2,1), scale=TRUE)
}
}
