% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier.R
\name{bezier_interp2}
\alias{bezier_interp2}
\title{Interpolate a Bézier Curve}
\usage{
bezier_interp2(coords, start, steps)
}
\arguments{
\item{coords}{"data.frame" or "list" containing "x" and "y" columns
corresponding to the points in cubic bezier, with \verb{3n} rows, where \code{n}
is the number of Bezier curves spliced together.}

\item{start}{numeric(2) x and y coordinates for the start point of the bezier}

\item{steps}{how many steps to use for each curve}
}
\description{
Interpolate a Bézier Curve
}
