% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{compare_svg}
\alias{compare_svg}
\alias{svg_samples}
\alias{compare_rsvg}
\title{Tools to Compare SVG Renderings}
\usage{
compare_svg(
  source = svg_samples(),
  target = tempfile(),
  ppi = getOption("svgchop.ppi", 125),
  width = 400,
  height = NA_real_,
  cols = 1,
  display = 2,
  timeout = 2,
  rsvg = FALSE,
  ...
)

svg_samples(pattern = "\\\\.svg$")

compare_rsvg(..., width = 400, display = 2, timeout = 2)
}
\arguments{
\item{target}{a file to write the html to.}

\item{width}{numeric(1L) if NA will use \code{height} and the aspect ratio from
the "viewBox" if present, or from the element extents if not.}

\item{height}{numeric(1L) if NA will use \code{height} and the aspect ratio from
the "viewBox" if present, or from the element extents if not.}

\item{cols}{integer(1L) how many columns to arrange the diptychs in (only for
\code{compare_svg}.}

\item{display}{numeric in 0:2, where 0 does not display, 1 opens the
generated HTML in a browser, and 2 (default) opens the generated HTML in a
browser, and after \code{timeout} seconds (enough time for browser to open)
deletes the files (to avoid cluttering drive during testing).}

\item{...}{additional arguments passed on to \code{\link[=vs_svg]{vs_svg()}} and \code{\link[=chop]{chop()}}}

\item{pattern}{character(1L) a regular expression to sub-select svgs from the
built-in samples.}

\item{character}{vector of paths or URLs for SVGs to compare.}
}
\value{
character(1L) the name of the file written to
}
\description{
\code{compare_svg} juxtaposes an SVG file to the corresponding \code{\link[=chop]{chop()}}ed and
rasterized (via \code{\link[=plot.svg_chopped]{plot.svg_chopped()}}) version.  \code{compare_rsvg} does the same,
except the reference SVG is rasterized with \code{\link[rsvg:rsvg_png]{rsvg::rsvg_png()}}.  For the
latter we also show the mean absolute difference for each pixel across
channels as a monochrome image.  Reference renderings are on the left.  Both
these functions default to the built-in svgs listable via \code{svg_samples}.
}
\examples{
\dontrun{
## These open browser instances
compare_svg()
compare_rsvg()
}
\donttest{
samples <- svg_samples()
scol <- ceiling(sqrt(length(samples)))
srow <- ceiling(length(samples) / scol)
svgs <- as.svg_chopped_list(lapply(samples, chop))
plot(svgs, mfcol=c(scol, srow), scale=TRUE)
}
}
