% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_xy_coords}
\alias{get_xy_coords}
\alias{get_fills}
\alias{get_strokes}
\title{Retrieve Basic Data From "svg_chopped" Objects}
\usage{
get_xy_coords(x)

get_fills(x)

get_strokes(x)
}
\arguments{
\item{x}{an "svg_chopped" or "svg_chopped_flat" object.}
}
\value{
a list or character vector of the same length as \code{x}.  See details.
}
\description{
Specialized retrieval functions designed to return a small but useful subset
of the data encoded in "svg_chopped" objects so you don't have to wade
through the complex structure of those objects.
}
\details{
"svg_chopped" objects are first \code{\link[=flatten]{flatten()}}ed and the data is retrieved from
each displayable element.

\code{get_xy_coords} returns the coordinates in a format compatible with
\code{\link[grDevices:xy.coords]{grDevices::xy.coords()}} (a list with x and y coordinates as numeric
vectors), and \code{get_fills} and \code{get_strokes} return a character vector with
colors processed by \code{\link[=approximate_color]{approximate_color()}}.
}
\examples{
svg <- chop(R_logo(), steps=3)
str(xy <- get_xy_coords(svg))
(fills <- get_fills(svg))
if(interactive()) {
  plot.new()
  ext <- attr(svg, "extents")
  plot.window(ext$x, rev(ext$y), asp=1)

  polypath(xy[[1]], col=fills[[1]], border=NA)
  polypath(xy[[2]], col=fills[[2]], border=NA)
}
}
\seealso{
\code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}, \code{\link[=approximate_color]{approximate_color()}},
\code{\link[=plot.svg_chopped]{plot.svg_chopped()}}.  @export
}
