% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_path}
\alias{parse_path}
\alias{parse_poly}
\title{Convert SVG Path to More Usable format}
\usage{
parse_path(x)

parse_poly(x)
}
\arguments{
\item{x}{a list representing a single SVG "path", which each element of the
list a property of the path.  The "d" property will be a list of "subpath"
S3 objects.}
}
\value{
a list with as many elements as there are sub-paths in the path "d"
property.  Each element is a "subpath" S3 object containing the path
commands and coordinates in a data frame, and all other path properties as
strings
}
\description{
For polygons there is only ever one sub-path.
}
