% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_paths}
\alias{parse_paths}
\title{Retrieve SVG Paths From File}
\usage{
parse_paths(file)
}
\arguments{
\item{file}{an SVG file}
}
\value{
an "svg_paths" S3 object, which is a list of "svg_path" objects with
the \code{x}, \code{y}, \code{width}, and \code{height} values of the outer SVG element
recorded in the "box" attribute.
}
\description{
Pull all paths out of SVG file and converts the "d" attribute into a path
command using only M, C, and L commands.  Only a subset of the path commands
are handled.
}
\seealso{
\code{\link[=interp_paths]{interp_paths()}}
}
