% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_svg}
\alias{parse_svg}
\title{Retrieve SVG Elements From File}
\usage{
parse_svg(file, steps = 10)
}
\arguments{
\item{file}{an SVG file}

\item{elements}{character the types of elements to parse, currently only
"path" and "polygon" are supported.}
}
\value{
an "svg_paths" S3 object, which is a list of "subpath" that have been
converted from their original SVG form to a line segments.  The \code{x}, \code{y},
\code{width}, and \code{height} values of the outer SVG element recorded in the "box"
attribute.
}
\description{
Pull all paths and polygons out of an SVG file and convert them to x-y
coordinates and L, M, and C SVG path commands corresponding to line segments.
Originally this was all built around SVG paths, so we're forcing polygons
through that pipeline even though we really don't need to.
}
\details{
If the document contains multiple SVG elements, only the first will be
parsed.  In the future will likely add the option to directly pass an \code{xml2}
node to allow iterating over multiple svgs.
}
\seealso{
\code{\link[=interp_paths]{interp_paths()}}
}
