% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{path_to_abs}
\alias{path_to_abs}
\title{Convert Path To Absolute Curve and Lines}
\usage{
path_to_abs(path)
}
\arguments{
\item{list}{of lists, each sub-lists contains the command type (e.g. M, C, L,
V, H, Z, m, c, l, v, h, z) at position one, and then a set of coordinates
as a numeric vector at position two.}

\item{x}{starting x coordinate}

\item{y}{starting y coordinate}
}
\value{
a list of lists similar to the input, except the only commands
therein will be M, C, L, and coordinates will be absolute
}
\description{
Recomputes any relative commands into their absolute equivalents,
replaces H and V commands with their L equivalents, explicitly closes paths
with L commands instead of Z.
}
\details{
Assumes first command is a "moveto" command.
}
