% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{R_logo}
\alias{R_logo}
\title{Return a Path to the SVG R Logo}
\usage{
R_logo(internal = TRUE)
}
\arguments{
\item{internal}{TRUE (default) whether to use the logo bundled with this
package, or FALSE to use the one in R's installation directory.}
}
\value{
character(1L) a path to R Logo SVG file.
}
\description{
For demo and examples.
}
