% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{str.svg_chopped}
\alias{str.svg_chopped}
\alias{str.svg_chopped_list}
\alias{str.svg_chopped_flat}
\alias{str.svg_chopped_list_flat}
\title{str Method "svg_chopped*" Objects}
\usage{
\method{str}{svg_chopped}(object, give.attr = FALSE, ...)

\method{str}{svg_chopped_list}(object, give.attr = FALSE, ...)

\method{str}{svg_chopped_flat}(object, give.attr = FALSE, ...)

\method{str}{svg_chopped_list_flat}(object, give.attr = FALSE, ...)
}
\arguments{
\item{object}{any \R object about which you want to have some
    information.}

\item{give.attr}{logical; if \code{TRUE} (default), show attributes
    as sub structures.}

\item{...}{potential further arguments (required for Method/Generic
    reasons).}
}
\value{
NULL, invisibly.
}
\description{
Exactly like \code{\link[utils:str]{utils::str()}}, except the \code{give.attr} parameter is set to FALSE
by default instead of TRUE as otherwise the output is overwhelming.  You can
set it explicitly to TRUE to show attributes.
}
\seealso{
\code{\link[=flatten]{flatten()}} to generate a flat version of the tree that may be
easier to subset.
}
