% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{styles_computed}
\alias{styles_computed}
\title{Report What Styles Are Computed}
\usage{
styles_computed()
}
\value{
character names of styles computed by \code{svgchop}
}
\description{
\code{svgchop} computes a small subset of style attributess based on element
attributes, CSS style sheets, and inheritance.  This function returns which
styles receive that treatment.
}
\examples{
styles_computed()
}
