% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{[.svg_chopped}
\alias{[.svg_chopped}
\alias{[.svg_chopped_flat}
\alias{[.svg_chopped_list}
\alias{[.svg_chopped_list_flat}
\title{Subset "svg_chopped_flat" Objects}
\usage{
\method{[}{svg_chopped}(x, i, ...)

\method{[}{svg_chopped_flat}(x, i, ...)

\method{[}{svg_chopped_list}(x, i, ...)

\method{[}{svg_chopped_list_flat}(x, i, ...)
}
\arguments{
\item{x}{an "svg_chopped" or related object}

\item{i}{mixed subsetting indices}
}
\description{
Versions of the base subsetting functions that do not drop attributes.  Among
other things, this allows us to subset complex objects and still benefit from
the inspection methods in the package.
}
