% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{svg_colors}
\alias{svg_colors}
\alias{svg_colors_all}
\title{SVG Color to Hex Code Mapping}
\usage{
svg_colors(colors)

svg_colors_all()
}
\arguments{
\item{colors}{character of color keywords (names).}
}
\value{
character vector named with color and with the 6 digit RGB hex code
as the value.  Unmatched color names return NA.
}
\description{
A list of the 147 known SVG colors from the
\href{https://www.w3.org/TR/SVG11/types.html#ColorKeywords}{W3C SVG 1.1. Spec}.
}
\examples{
svg_colors(c('green', 'plum', 'NOT A REAL COLOR'))
length(svg_colors_all()) # 147 named colors
}
