% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{svg_gallery}
\alias{svg_gallery}
\alias{svg_samples}
\title{Compare SVGs to Their "Chopped" Renderings}
\usage{
svg_gallery(
  source = svg_samples(),
  target = tempfile(),
  ppi = getOption("svgchop.ppi", 125),
  width = 400,
  height = NA_real_,
  cols = 1,
  display = 2,
  timeout = 2,
  ...
)

svg_samples(pattern = "\\\\.svg$")
}
\arguments{
\item{source}{character vector of paths or URLs for SVGs to compile into
single HTML file.}

\item{target}{a file to write the html to.}

\item{width}{numeric(1L) if NA will use \code{height} and the aspect ratio from
the "viewBox" if present, or from the element extents if not.}

\item{height}{numeric(1L) if NA will use \code{height} and the aspect ratio from
the "viewBox" if present, or from the element extents if not.}

\item{cols}{integer(1L) how many columns to arrange the diptychs in.}

\item{display}{numeric in 0:2, where 0 does not display, 1 opens the
generated HTML in a browser, and 2 (default) opens the generated HTML in a
browser, and after \code{timeout} seconds (enough time for browser to open)
deletes the files (to avoid cluttering drive during testing).}

\item{...}{additional arguments passed on to \code{\link[=chop]{chop()}}}
}
\value{
character(1L) the name of the file written to
}
\description{
Create an HTML page of SVG and their corresponding \code{\link[=chop]{chop()}}ed and rasterized
(png) counterparts juxtaposed into diptychs for comparison.  Viewport and
viewbox parameters may be manipulated to ensure the images fit into boxes of
width controlled by the \code{width} parameter.  Original SVG will be on the left,
with the rasterized counterpart on the right.
}
\examples{
\dontrun{
svg_gallery()   # opens a browser instance
}
\donttest{
samples <- svg_samples()
scol <- ceiling(sqrt(length(samples)))
srow <- ceiling(length(samples) / scol)
svgs <- as.svg_chopped_list(lapply(samples, chop))
plot(svgs, mfcol=c(scol, srow), scale=TRUE)
}
}
