% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgchop.R
\docType{package}
\name{svgchop}
\alias{svgchop}
\title{Approximate SVG Elements With Line Segments}
\description{
Parse SVG elements and approximate them with straight line segments.  Only a
small set of SVG syntax and commands are recognized.  The parsing is done
with regex so don't expect it to be robust.  The code is slow and hacky.
}
\details{
This is package is experimental, lightly tested, and not intended for
production use.
}
