% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_coords}
\alias{transform_coords}
\title{Compute and Apply SVG Transforms to Coordinates}
\usage{
transform_coords(x)
}
\arguments{
\item{x}{an "svg_chopped" object as produced by \code{\link[=parse_svg]{parse_svg()}}.}
}
\value{
x, with coordinates transformed.
}
\description{
SVG transforms should be applied recursively to nested elements.  We
implement this by recursively collecting the "transform" XML attributes,
parsing them into transformation matrices, and ultimately applying the
accumulated transformation matrices to the terminal SVG elements.
}
\seealso{
\code{\link[=parse_svg]{parse_svg()}}
}
