% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{compute_display_params}
\alias{compute_display_params}
\title{Compute Display Parameters for Device}
\usage{
compute_display_params(
  x,
  pin = par("pin"),
  ppi = getOption("svgchop.ppi", 125),
  scale = FALSE,
  center = TRUE
)
}
\arguments{
\item{x}{an "svg_chopped" object}

\item{pin}{numeric length 2 width and height of the plot area in inches.}

\item{ppi}{numeric device resolution in pixels per inch; if rendered SVGs
look larger or smaller than in your browser you may need to adjust this
setting.}

\item{scale}{TRUE (currently unsupported) or FALSE (default) whether units
should be scaled to fit the viewbox in the display.  Aspect ratio will be
preserved.}
}
\value{
a list containing elements:
\itemize{
\item "plot.lim": a list with the x and y plot limits in user coordinates.
\item "asp": numeric(1) the height/width ratio of each pixel in the plot (not the
actual aspect ratio of the plot).
\item "uppi": numeric(1) the user-pixels per inch.
}
}
\description{
Computes how the dimensions of the display device in user coordinates, the
aspect ratio of user coordinates relative to device coordinates, and the user
coordinates "ppi" (this is taken to be the higher of the X or Y ppi, which
may not be the same if the aspect ratio is not 1).
}
\details{
This information can be used by a rendering function such as
\code{\link[=plot.svg_chopped]{plot.svg_chopped()}} to scale and position the output.
}
