% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{flatten}
\alias{flatten}
\alias{flatten.default}
\alias{flatten.svg_chopped}
\alias{flatten.svg_chopped_list}
\title{Simplify "svg_chopped" Structure}
\usage{
flatten(x, ...)

\method{flatten}{default}(x, ...)

\method{flatten}{svg_chopped}(x, ...)

\method{flatten}{svg_chopped_list}(x, ...)
}
\arguments{
\item{x}{an object to flatten}
}
\value{
the object flattened
}
\description{
\code{flatten} collapses "svg_chopped" recursive structure by placing all
non-"hidden" terminal nodes into a one level list.  Flattened
"svg_chopped_list" objects will retain distinct elements for each
"svg_chopped" contained therein, so an "svg_chopped_list_flat" object will
have two levels: one for the "svg_chopped_flat" objects, and one for the
visible terminal leaves of those.
}
\details{
When flattening terminal leaves are retrieved via depth-first recursion into
a single level list for each "svg_chopped" object.  Terminal elements
defined inside "defs" will be hidden.  Attributes for the "svg_chopped",
"svg_chopped_list", and terminal nodes are retained.

For convenience the flat list is named with the numeric index of the element
and the svg element name.  The underlying recursive list is unnamed so we
use the name to succinctly display key information about the object when
examined with \code{\link[utils:str]{utils::str()}} (see example).
}
\examples{
svg <- process_svg(file.path(R.home(), 'doc', 'html', 'Rlogo.svg'))
str(flatten(svg), give.attr=FALSE)   # attributes may be overwhelming
}
