% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_path}
\alias{parse_path}
\title{Parse "d" Path Command}
\usage{
parse_path(x)
}
\arguments{
\item{x}{character a vector of the contents of the d attribute of SVG paths}
}
\value{
a list of of length equal to \code{x}'s, with each element a list
containing as many data frames  as there are sub-paths in the corresponding
\code{x} element, with each data frame containing a column with commands in
\code{c("M","L","C")}.
}
\description{
Convert "d" path attribute into a more usable format containing only "M",
"C", and "L" commands.
}
