% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_path}
\alias{parse_path}
\title{Convert SVG Path to More Usable format}
\usage{
parse_path(x)
}
\arguments{
\item{x}{a list representing a single SVG "path", which each element of the
list a property of the path.  It may be useful to ensure paths have an id
attribute to keep track of where each sub-path produced by this function
came from.}
}
\value{
a list with as many elements as there are sub-paths in the path "d"
property.  Each element is a "subpath" S3 object containing the path
commands and coordinates in a data frame, and all other path properties as
strings
}
\description{
Convert SVG Path to More Usable format
}
