% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{parse_path}
\alias{parse_path}
\title{Convert SVG Path to Line Segments}
\usage{
parse_path(x, steps = 20)
}
\arguments{
\item{x}{a path SVG node.}

\item{steps}{positive integer(1), how many line segments to use to
approximate Bézier curves or elliptical arcs.  For arcs, it is the number
of steps for a complete ellipse, so for partial ellipses fewer steps will
be used.}
}
\value{
numeric matrix, 2 x n containing the x and y coordinates of the
endpoints of the n - 1 concatenated line segments that approximate the path
described by the "d" attribute of the path SVG element \code{x}.  If there is
more than one sub-path, the starting column of sub-paths following the
first will be stored as the "starts" attribute of the matrix.
}
\description{
Parses the "d" path attribute into X-Y coordinates of line segments collected
into sub-paths.  Sub-paths are designated by "M" or "m" commands embedded in
the path command.  Bézier curves and paths are interpolated.
}
