% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.svg_chopped}
\alias{plot.svg_chopped}
\title{Plot Method for "svg_chopped" Objects}
\usage{
\method{plot}{svg_chopped}(x, ...)
}
\arguments{
\item{x}{an "svg_chopped" object}

\item{...}{unused}
}
\value{
\code{x}, invisibly
}
\description{
Display an SVG plot object to the graphical device.  Renders closed polygons
with \code{\link[graphics:polypath]{graphics::polypath()}} and open ones with \code{\link[graphics:lines]{graphics::lines()}}.  Will set
the viewport to the SVG viewBox coordinates if they exist, and to the SVG
extent if they don't.  Aspect ratio is fixed so that user coordinates are 1:1
in x and y directions.
}
\details{
Nested SVGs viewports will not be respected and their contents will be drawn
in the outer viewport coordinates.

Stroke widths are computed under the assumption that 1 "lwd" == 1/96th of an
inch, and may look bad on devices that don't support "lwd" values less than
one.
}
