% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{process_svg}
\alias{process_svg}
\title{Convert SVG Elements to Polygons}
\usage{
process_svg(file, steps = 10)
}
\arguments{
\item{file}{an HTML or other XML based text file containing SVG elements.}
}
\value{
an "svg_chopped_list" S3 object, which is a list of "svg_chopped"
objects.  Each "svg_chopped" object represents an SVG viewport the
dimensions of which are recorded in the "box" attribute.  "svg_chopped"
objects are recursive lists with \verb{2 x n} numeric matrices as terminal
leaves.  The matrices contain the X-Y coordinates of the ordered \code{n}
endpoints of the \code{n - 1} line segments that the polygon representation of
the SVG elements comprise.
}
\description{
Parse and convert SVG elements into polygons.  SVG transforms are applied to
the polygon coordinates, and SVG presentation attributes are computed from
style sheets, inline styles and attributes, and are attached as the
"style-computed" R attribute.
}
\details{
Currently only SVG path, rect, and polygon elements are supported.  Only some
transforms are implemented.  CSS support is likely to be particularly fragile
as the CSS parsing is regex based and only simple ASCII-only class and id
selectors are supported.
}
